% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{add_join_helpers}
\alias{add_join_helpers}
\title{add_join_helpers}
\usage{
add_join_helpers(df, sched = nfl_query("select * from nflschedule"))
}
\arguments{
\item{df}{a dataframe to add the columns to}

\item{sched}{a dataframe of nfl schedule data}
}
\description{
function to add season and week columns to assist in creating
average tables
}

