% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_date}
\alias{convert_to_date}
\title{convert_to_date}
\usage{
convert_to_date(df, colname = "date")
}
\arguments{
\item{df}{dataframe to fix date for.}

\item{colname}{Name of date column. Defaults to 'date'.}
}
\value{
Returns dataframe with date column fixed.
}
\description{
Utility function to change integer to date.
}

