% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nfl_def_profiles.R
\name{create_nfl_def_profiles}
\alias{create_nfl_def_profiles}
\title{create_nfl_def_profiles}
\usage{
create_nfl_def_profiles(df, window_size = 3)
}
\arguments{
\item{df}{Dataframe to create profile from}

\item{window_size}{Size of rolling window. Defaults to 3.}
}
\value{
Returns list of profiles.
}
\description{
Function to profiles for nfl defenses.
}

