% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nfl_off_profiles.R
\name{create_nfl_off_profiles}
\alias{create_nfl_off_profiles}
\title{create_nfl_off_profiles}
\usage{
create_nfl_off_profiles(df, window_size = 3)
}
\arguments{
\item{df}{Dataframe to create profiles from.}

\item{window_size}{Size of rolling window. Defaults to 3.}
}
\value{
Returns nested list of lists containing profile, projection, and opponent.
}
\description{
Function to create profiles for NFL players.
}

