% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profiles.R
\name{create_profiles}
\alias{create_profiles}
\title{create_profiles}
\usage{
create_profiles(df, window_size = 3)
}
\arguments{
\item{df}{Dataframe to create profiles from.}

\item{window_size}{Size of rolling window. Defaults to 3.}
}
\value{
Returns nested list of lists containing profile and projection.
}
\description{
Function to create profiles for players.
}

