% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfs_insert}
\alias{dfs_insert}
\title{dfs_insert}
\usage{
dfs_insert(db_path = Sys.getenv("DFS_DB"), table, df, ...)
}
\arguments{
\item{db_path}{Path to sqlite database. Defaults to DFS_DB variable in .Renviron.}

\item{table}{Name of table to write data to.}

\item{df}{The dataframe to write.}

\item{...}{additional parameters for DBI::dbWriteTAble}
}
\description{
Function to insert data into dfs db.
}

