% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_def}
\alias{fill_def}
\title{fill_def}
\usage{
fill_def(df, full_sched = nfl_query("select * from vegas"))
}
\arguments{
\item{df}{dataframe to fill}

\item{full_sched}{dataframe showing all of team's opponents}

\item{sched}{dataframe used to add year and week columns}
}
\description{
if a position registered no stats, fill row
}

