% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_defavgs.R
\name{generate_defavgs}
\alias{generate_defavgs}
\title{generate_defavgs}
\usage{
generate_defavgs(qbdef = nfl_query("select * from qbdef"),
  rbdef = nfl_query("select * from rbdef"),
  wrdef = nfl_query("select * from wrdef"),
  tedef = nfl_query("select * from tedef"), target_date = "2002-08-31")
}
\arguments{
\item{qbdef}{dataframe of quarterback defense}

\item{rbdef}{dataframe of runningback defense}

\item{wrdef}{dataframe of wide receiver defence}

\item{tedef}{dataframe of te_def}

\item{target_date}{date to determine what data to generate defensive statistics
from}
}
\description{
\code{generate_defstats} Generates league average defensive statistics for
each position
}

