% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_defstats.R
\name{generate_defstats}
\alias{generate_defstats}
\title{generate_defstats}
\usage{
generate_defstats(target_date)
}
\arguments{
\item{target_date}{date to determine what data to generate defensive statistics
from}
}
\description{
\code{generate_defstats} Generates defensive statistics from box score data and
inserts into appropriate database table. Will be utilized after completion of
week to update defense tables.
}

