% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boxscore_links.R
\name{get_boxscore_links}
\alias{get_boxscore_links}
\title{get_boxscore_links}
\usage{
get_boxscore_links(year, week = 1:17)
}
\arguments{
\item{year}{Season to get links from.}

\item{week}{Week interested in. Defaults to all weeks(1 - 17).}
}
\value{
Returns list of links.
}
\description{
Function to get links for boxscores from the schedule page.
}

