% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boxscore_stats.R
\name{get_nfl_boxscore_stats}
\alias{get_nfl_boxscore_stats}
\title{get_nflboxscore_stats}
\usage{
get_nfl_boxscore_stats(url, insert = FALSE)
}
\arguments{
\item{url}{Url of boxscore.}

\item{insert}{Should the function insert the data into the nfl_offense and
nfl_defense tables? Defaults to FALSE.}
}
\value{
Returns list containing offensive stats (off_stats) and defensive
stats (def_stats)
}
\description{
Function to scrape NFL offense stats and snap counts tables.
Generates defensive tables from offense stats.
}

