% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boxscore_stats.R
\name{get_nfl_positions}
\alias{get_nfl_positions}
\title{get_nfl_positions}
\usage{
get_nfl_positions(links)
}
\arguments{
\item{links}{Player page urls}
}
\value{
Returns positions.
}
\description{
Get NFL player's position if snap counts table is unavailable.
If player has two listed positions, first one listed will be taken.
}

