% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boxscore_stats.R
\name{get_positions}
\alias{get_positions}
\title{get_positions}
\usage{
get_positions(links)
}
\arguments{
\item{links}{Player page urls}
}
\value{
Returns positions.
}
\description{
Get player's position if snap counts table is unavailable.
If player has two listed positions, first one listed will be taken.
}

