% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_week_dates.R
\name{get_week_dates}
\alias{get_week_dates}
\title{get_week_dates}
\usage{
get_week_dates(year, insert = FALSE)
}
\arguments{
\item{year}{Season to get data for.}

\item{insert}{Should the function insert data into nfl_week_date table?
Defaults to FALSE.}
}
\value{
Returns dataframe of dates for each week.
}
\description{
Function to get the dates for each week.
}

