% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nba_insert}
\alias{nba_insert}
\title{nba_insert}
\usage{
nba_insert(dataframe, table)
}
\arguments{
\item{dataframe}{dataframe to insert}

\item{table}{the name of the table to insert into}
}
\description{
insert data to nba database
}

