% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nba_query}
\alias{nba_query}
\title{nba_query}
\usage{
nba_query(query)
}
\arguments{
\item{query}{query to send to database}
}
\value{
results of query in a dataframe
}
\description{
function to make queries to nba schema
}

