% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nfl_insert}
\alias{nfl_insert}
\title{nfl_insert}
\usage{
nfl_insert(db_name = Sys.getenv("SQLITE_DB"), table, df)
}
\arguments{
\item{table}{Name of table to write data to.}

\item{df}{The dataframe to write.}

\item{db_path}{Path to sqlite database. Defaults to SQLITE_DB variable in .Renviron.}
}
\description{
\code{nfl_insert} a utility function to insert data into nfl schema
}

