% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_optimize.R
\name{nfl_optimize}
\alias{nfl_optimize}
\title{nfl_optimize}
\usage{
nfl_optimize(df, platform)
}
\arguments{
\item{df}{dataframe containing player projections, salaries, positions}

\item{platform}{'dk' or 'fd' depending on platform}
}
\value{
dataframe of chosen lineup
}
\description{
build optimzed line up using simplex algorithm
}

