% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{oppdate_fixer}
\alias{oppdate_fixer}
\title{oppdate_fixer}
\usage{
oppdate_fixer(df, v)
}
\arguments{
\item{v}{vegas dataframe}

\item{def}{dataframe of defensive stats}
}
\value{
returns updated dataframe
}
\description{
updates opponent and date for defenses
}

