% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_qb.R
\name{predict_qb}
\alias{predict_qb}
\title{predict_qb}
\usage{
predict_qb(qb_dflist, price)
}
\arguments{
\item{qb_dflist}{list of qb dataframes (one ,two, three week based)}

\item{price}{df of player prices}
}
\value{
returns dataframe of predictions
}
\description{
make predictions for qb
}

