% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_rb.R
\name{predict_rb}
\alias{predict_rb}
\title{predict_rb}
\usage{
predict_rb(rb_dflist, price)
}
\arguments{
\item{rb_dflist}{list of rb dataframes (one ,two, three week based)}

\item{price}{df of player prices}
}
\value{
returns dataframe of predictions
}
\description{
make predictions for rb
}

