% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_te.R
\name{predict_te}
\alias{predict_te}
\title{predict_te}
\usage{
predict_te(te_dflist, price)
}
\arguments{
\item{te_dflist}{list of te dataframes (one ,two, three week based)}

\item{price}{df of te prices}
}
\value{
returns dataframe of predictions
}
\description{
make predictions for te
}

