% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{roll_n}
\alias{roll_n}
\title{roll_n}
\usage{
roll_n(df, n, pts_name, num_start, weeks, def = FALSE)
}
\arguments{
\item{df}{dataframe to find rolling average}

\item{n}{number of weeks to look base prediction on}

\item{pts_name}{name of column that contains fantasy points}

\item{num_start}{name of column that starts stats}

\item{def}{boolean that indicates if df is a defensive stats dataframe,
default is FALSE}
}
\description{
finds rolling average for previous n weeks, adds actual stat columns
}

