% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/salary_parser.R
\name{salary_parser}
\alias{salary_parser}
\title{salary_parser}
\usage{
salary_parser(platform, week)
}
\arguments{
\item{platform}{string indicating 'DK' or 'FD'}

\item{week}{the week you are making predictions for}
}
\value{
dataframe of parsed df to be used by other dfstoolkit functions
}
\description{
parses salary data from DraftKings or FanDuel
}

