% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_one}
\alias{shift_one}
\title{shift_one}
\usage{
shift_one(df, pts_name, num_start, def = FALSE)
}
\arguments{
\item{df}{dataframe to offset}

\item{pts_name}{name of column that contains fantasy points}

\item{num_start}{name of column that starts stats}

\item{def}{boolean that indicates if df is a defensive stats dataframe,
default is FALSE}
}
\description{
offsets data by one week, adds actual stat columns
}

