% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqlite2posix}
\alias{sqlite2posix}
\title{sqlite2posix}
\usage{
sqlite2posix(x)
}
\arguments{
\item{x}{numeric value to change to POSIXct}
}
\value{
Returns dataframe with date column fixed.
}
\description{
Utility function to change integer to date.
}

