% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{nfl_insert}
\alias{nfl_insert}
\title{nfl_insert}
\usage{
nfl_insert(dataframe, table)
}
\arguments{
\item{dataframe}{a dataframe that contains the data to be inserted into table}

\item{table}{the name of the table to insert the dataframe into}
}
\description{
\code{nfl_insert} a utility function to insert data into nfl schema
}

