% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nfl_query}
\alias{nfl_query}
\title{nfl_query}
\usage{
nfl_query(query, target_date = "2002-09-05")
}
\arguments{
\item{query}{query to send to database}

\item{target_date}{query will only get data at or after this date. Defaults
to start of 2002 NFL season.}
}
\value{
results of query in a dataframe
}
\description{
\code{nfl_query} a utility function to get data from nfl schema
}

