% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nfl_query}
\alias{nfl_query}
\title{dfs_query}
\usage{
nfl_query(db_path = Sys.getenv("DFS_DB"), query)
}
\arguments{
\item{db_path}{Path to sqlite database. Defaults to DFS_DB variable in .Renviron.}

\item{query}{query to send to database}
}
\value{
results of query in a dataframe
}
\description{
Function to get data from dfs db.
}

