% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oppdate_fixer}
\alias{oppdate_fixer}
\title{oppdate_fixer}
\usage{
oppdate_fixer(df, v, bye_teams)
}
\arguments{
\item{v}{vegas dataframe}

\item{bye_teams}{teams that are on bye}

\item{def}{dataframe of defensive stats}
}
\value{
returns updated dataframe
}
\description{
updates opponent and date for defenses
}

