% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_wr.R
\name{predict_wr}
\alias{predict_wr}
\title{predict_wr}
\usage{
predict_wr(wr_dflist, price)
}
\arguments{
\item{wr_dflist}{list of wr dataframes (one ,two, three week based)}

\item{price}{df of wr prices}
}
\value{
returns dataframe
}
\description{
make predictions for wr
}

