% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roll_n}
\alias{roll_n}
\title{roll_n}
\usage{
roll_n(df, n, pts_name, num_name)
}
\arguments{
\item{df}{dataframe to find rolling average}

\item{n}{number of weeks}

\item{pts_name}{name of column that contains fantasy points}

\item{num_start}{name of column that starts stats}
}
\value{
returns rolled dataframe
}
\description{
find means for players and defenses
}

